+++ shoebox-ejb/src/main/java/parametrageCoop/serviceParamCoopLocal.java
@@ -127,7 +127,7 @@ public interface serviceParamCoopLocal {
 
 
     //***************************autres*********************************//
-    public List<Object[]> rechercheStockProduit(Produit produit, int grade, Cooperative coop,DefinitionPeriode periode);
+    public List<Object[]> rechercheStockProduit(Produit produit, Long grade, Cooperative coop,DefinitionPeriode periode);
 
     public List<Object[]> rechercheStockProduitIntrant(Produit produit, Cooperative coop, DefinitionPeriode periode);
